+++ apps/server/src/modules/board/controller/dto/card/index.ts
@@ -1,3 +1,4 @@
+export * from './any-content-element.response';
 export * from './card-ids.params';
 export * from './card-list.response';
 export * from './card.response';
